Changed callbacks for new elements to Promises
I felt like there was a super easy chance that the user would fall into Callback Hell.  So I converted them over to promises to prevent this.  In Jest there was an issue and we have to make them async/await functions.  This was actually intended behavior because it was rendering before the promise could execute.  You need to ensure that you are rendering in a new then statement to ensure all the needed actions have executed.