Roadblock fixed
I ran into a road block where I was importing abstract class into the new class and the new class into abstract file.  This created a Stack Overflow and returned null as an import.  I needed to rewrite all of the code for the Markdown classes.  This is done and the tests have also been rewritten.  I also rewrote the Badge class for a better API flow.  Tests all updated and setup.  I also created a test and dev setup inside the project.json